+++ packages/xo-web/src/common/intl/messages.js
@@ -346,6 +346,9 @@ const messages = {
   missingSnapshotRetention:
     'The Rolling Snapshot mode requires snapshot retention to be higher than 0!',
   retentionNeeded: 'One of the retentions needs to be higher than 0!',
+  createRemoteMessage:
+    'No remotes found, please click on the remotes settings button to create one!',
+  remotesSettings: 'Remotes settings',
   scheduleAdd: 'Add a schedule',
   scheduleDelete: 'Delete',
   scheduleRun: 'Run schedule',